code/requirements.txt
#torch==1.13.0
#torchvision #==0.5.0
#pytorch-lightning==0.7.5
numpy
matplotlib
scikit-image
scipy
torchsummary
pyyaml
#tensorflow==1.15.2
tensorboard #==1.15
attrdict
opencv-python
moviepy
orderedattrdict
imageio
kornia
jupyter
test_tube #==0.7.5
